insulator_yolo/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "insulator-yolo"
version = "0.1.0"
description = "Reusable YOLO training pipeline for insulator defect detection."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0",
  "Pillow>=10.0",
  "pytest>=9.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
